app/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.1.0
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn>=1.0.0
joblib>=1.0.0